Refinements (#30)
* Refactor dynamic graph traits and aliases

* Enable Wshadow and apply shadow-related warning fixes

* Fix warning-cleanup fallout and sanitizer test discovery

* Update dynamic_graph traits documentation

* Fix GCC -Wshadow collisions with graph CPO names

* Specialize edge_list descriptor for void edge values

* Fix clang-cl build, benchmark flags, and topological sort stack overflow

- cmake: guard /MP flag to cl.exe only (clang-cl doesn't support it)
- cmake: add clang-cl-specific warning suppressions (-Wno-unknown-attributes,
  -Wno-unused-parameter/variable, propagate CLANG_WARNINGS -Wno-* flags)
- benchmark: add required 's' suffix to --benchmark_min_time (v1.9.1 API)
- topological_sort: convert recursive DFS to iterative to fix stack overflow
  on deep graphs (e.g. 10000-vertex path DAG in BM_TopoSort benchmark)
- Broad warning cleanup across views, algorithms, generators, io, tests,
  and examples for clang-cl compatibility (unused params, unqualified std::move/forward, etc.)

* fix: take uid by value in basic_incidence/basic_out_incidence to suppress GCC -Wdangling-reference

GCC 15 incorrectly flags the uid parameter in basic_incidence() and
basic_out_incidence() as a dangling reference when used in a range-for
structured binding. Since vertex_id_t is always a cheap integer or map
key type, passing by value is correct and suppresses the false positive.

+++ tests/adj_list/cpo/test_source_id_cpo.cpp
@@ -166,7 +166,7 @@ TEST_CASE("source_id(g,uv) - native edge member function", "[source_id][cpo][mem
   SECTION("Native edge member function is called") {
     std::vector<int> sources;
     for (auto e : edges(g.adjacency_list, v0)) {
-      sources.push_back(static_cast<int>(source_id(g.adjacency_list, e)));
+      sources.push_back(source_id(g.adjacency_list, e));
     }
 
     // Should use CustomEdge::source_id() which returns source * 100
